requirements.txt
autograd>=1.3
matplotlib==3.0.3
numpy>=1.16.2
pyMKL==0.0.3
scipy==1.2.1
scikit-image==0.16.2
PyYAML
PyQt5
pinyin
keyboard